SIPフレームをパースするcom層(今はまだトンネル)を追加した。
