+++ doc/設計.md
@@ -1,3 +1,10 @@
+# レイヤ構造
+```
+drv.py # IF層、socketによるパケット送受信
+└com.py # SIPドライバ層、パケット⇔フレーム相互変換
+　└～
+```
+
 # event.py
 - ブロッキング機能を持つqueueにより、イベントキューを実現する
 - 優先度付きキューにより、イベント発行時のフック起動順序を規定する
