+++ src/event.py
@@ -33,6 +33,7 @@ def exec():
     heapq.heapify(event_hook)
     while event_hook:
         _, h = heapq.heappop(event_hook)
+        # print(event_id)
         h(params)
